Move channel tests to test_core_gates.py
