+++ src/qibo/tests_new/test_core_gates.py
@@ -7,9 +7,21 @@ from qibo.config import raise_error
 
 
 def random_state(nqubits):
+    """Generates a random normalized state vector as numpy array."""
     nstates = 2 ** nqubits
     initial_state = np.random.random(nstates) + 1j * np.random.random(nstates)
     return initial_state / np.sqrt((np.abs(initial_state) ** 2).sum())
+
+
+def random_density_matrix(nqubits):
+    """Generates a random normalized density matrix."""
+    shape = 2 * (2 ** nqubits,)
+    m = np.random.random(shape) + 1j * np.random.random(shape)
+    rho = (m + m.T.conj()) / 2.0
+    # Normalize
+    ids = np.arange(2 ** nqubits)
+    rho[ids, ids] = rho[ids, ids] / np.trace(rho)
+    return rho
 
 
 def apply_gates(gatelist, nqubits=None, initial_state=None):
@@ -462,4 +474,156 @@ def test_callback_gate_errors():
         gate.construct_unitary()
 
 
-# TODO: Test channels
+def test_general_channel(backend):
+    original_backend = qibo.get_backend()
+    qibo.set_backend(backend)
+    from qibo import K
+    a1 = np.sqrt(0.4) * np.array([[0, 1], [1, 0]])
+    a2 = np.sqrt(0.6) * np.array([[1, 0, 0, 0], [0, 1, 0, 0],
+                                  [0, 0, 0, 1], [0, 0, 1, 0]])
+    a1, a2 = K.cast(a1), K.cast(a2)
+    initial_rho = random_density_matrix(2)
+    gate = gates.KrausChannel([((1,), a1), ((0, 1), a2)])
+    assert gate.target_qubits == (0, 1)
+    final_rho = gate(np.copy(initial_rho))
+    m1 = np.kron(np.eye(2), np.array(a1))
+    m2 = np.array(a2)
+    target_rho = (m1.dot(initial_rho).dot(m1.conj().T) +
+                  m2.dot(initial_rho).dot(m2.conj().T))
+    np.testing.assert_allclose(final_rho, target_rho)
+    qibo.set_backend(original_backend)
+
+
+def test_unitary_channel(backend):
+    original_backend = qibo.get_backend()
+    qibo.set_backend(backend)
+
+    a1 = np.array([[0, 1], [1, 0]])
+    a2 = np.array([[1, 0, 0, 0], [0, 1, 0, 0], [0, 0, 0, 1], [0, 0, 1, 0]])
+    probs = [0.4, 0.3]
+    matrices = [((0,), a1), ((2, 3), a2)]
+    initial_state = random_density_matrix(4)
+    gate = gates.UnitaryChannel(probs, matrices)
+    gate.density_matrix = True
+    final_state = gate(np.copy(initial_state))
+
+    eye = np.eye(2)
+    ma1 = np.kron(np.kron(a1, eye), np.kron(eye, eye))
+    ma2 = np.kron(np.kron(eye, eye), a2)
+    target_state = (0.3 * initial_state
+                    + 0.4 * ma1.dot(initial_state.dot(ma1))
+                    + 0.3 * ma2.dot(initial_state.dot(ma2)))
+    np.testing.assert_allclose(final_state, target_state)
+    qibo.set_backend(original_backend)
+
+
+def test_unitary_channel_errors():
+    """Check errors raised by ``gates.UnitaryChannel``."""
+    a1 = np.array([[0, 1], [1, 0]])
+    a2 = np.array([[1, 0, 0, 0], [0, 1, 0, 0], [0, 0, 0, 1], [0, 0, 1, 0]])
+    probs = [0.4, 0.3]
+    matrices = [((0,), a1), ((2, 3), a2)]
+    # Invalid probability length
+    with pytest.raises(ValueError):
+        gate = gates.UnitaryChannel([0.1, 0.3, 0.2], matrices)
+    # Probability > 1
+    with pytest.raises(ValueError):
+        gate = gates.UnitaryChannel([1.1, 0.2], matrices)
+    # Probability sum = 0
+    with pytest.raises(ValueError):
+        gate = gates.UnitaryChannel([0.0, 0.0], matrices)
+
+
+def test_pauli_noise_channel(backend):
+    original_backend = qibo.get_backend()
+    qibo.set_backend(backend)
+    initial_rho = random_density_matrix(2)
+    gate = gates.PauliNoiseChannel(1, px=0.3)
+    gate.density_matrix = True
+    final_rho = gate(np.copy(initial_rho))
+    gate = gates.X(1)
+    gate.density_matrix = True
+    target_rho = 0.3 * gate(np.copy(initial_rho))
+    target_rho += 0.7 * initial_rho
+    np.testing.assert_allclose(final_rho, target_rho)
+    qibo.set_backend(original_backend)
+
+
+def test_reset_channel(backend):
+    original_backend = qibo.get_backend()
+    qibo.set_backend(backend)
+    initial_rho = random_density_matrix(3)
+    gate = gates.ResetChannel(0, p0=0.2, p1=0.2)
+    gate.density_matrix = True
+    final_rho = gate(np.copy(initial_rho))
+
+    dtype = initial_rho.dtype
+    collapsed_rho = np.copy(initial_rho).reshape(6 * (2,))
+    collapsed_rho[0, :, :, 1, :, :] = np.zeros(4 * (2,), dtype=dtype)
+    collapsed_rho[1, :, :, 0, :, :] = np.zeros(4 * (2,), dtype=dtype)
+    collapsed_rho[1, :, :, 1, :, :] = np.zeros(4 * (2,), dtype=dtype)
+    collapsed_rho = collapsed_rho.reshape((8, 8))
+    collapsed_rho /= np.trace(collapsed_rho)
+    mx = np.kron(np.array([[0, 1], [1, 0]]), np.eye(4))
+    flipped_rho = mx.dot(collapsed_rho.dot(mx))
+    target_rho = 0.6 * initial_rho + 0.2 * (collapsed_rho + flipped_rho)
+    np.testing.assert_allclose(final_rho, target_rho)
+    qibo.set_backend(original_backend)
+
+
+@pytest.mark.parametrize("t1,t2,time,excpop",
+                         [(0.8, 0.5, 1.0, 0.4), (0.5, 0.8, 1.0, 0.4)])
+def test_thermal_relaxation_channel(backend, t1, t2, time, excpop):
+    """Check ``gates.ThermalRelaxationChannel`` on a 3-qubit random density matrix."""
+    original_backend = qibo.get_backend()
+    qibo.set_backend(backend)
+    initial_rho = random_density_matrix(3)
+    gate = gates.ThermalRelaxationChannel(0, t1, t2, time=time,
+        excited_population=excpop)
+    gate.density_matrix = True
+    final_rho = gate(np.copy(initial_rho))
+
+    exp, p0, p1 = gate.calculate_probabilities(t1, t2, time, excpop)
+    if t2 > t1:
+        matrix = np.diag([1 - p1, p0, p1, 1 - p0])
+        matrix[0, -1], matrix[-1, 0] = exp, exp
+        matrix = matrix.reshape(4 * (2,))
+        # Apply matrix using Eq. (3.28) from arXiv:1111.6950
+        target_rho = np.copy(initial_rho).reshape(6 * (2,))
+        target_rho = np.einsum("abcd,aJKcjk->bJKdjk", matrix, target_rho)
+        target_rho = target_rho.reshape(initial_rho.shape)
+    else:
+        pz = exp
+        pi = 1 - pz - p0 - p1
+        dtype = initial_rho.dtype
+        collapsed_rho = np.copy(initial_rho).reshape(6 * (2,))
+        collapsed_rho[0, :, :, 1, :, :] = np.zeros(4 * (2,), dtype=dtype)
+        collapsed_rho[1, :, :, 0, :, :] = np.zeros(4 * (2,), dtype=dtype)
+        collapsed_rho[1, :, :, 1, :, :] = np.zeros(4 * (2,), dtype=dtype)
+        collapsed_rho = collapsed_rho.reshape((8, 8))
+        collapsed_rho /= np.trace(collapsed_rho)
+        mx = np.kron(np.array([[0, 1], [1, 0]]), np.eye(4))
+        mz = np.kron(np.array([[1, 0], [0, -1]]), np.eye(4))
+        z_rho = mz.dot(initial_rho.dot(mz))
+        flipped_rho = mx.dot(collapsed_rho.dot(mx))
+        target_rho = (pi * initial_rho + pz * z_rho + p0 * collapsed_rho +
+                      p1 * flipped_rho)
+    np.testing.assert_allclose(final_rho, target_rho)
+    # Try to apply to state vector if t1 < t2
+    if t1 < t2:
+        with pytest.raises(ValueError):
+            gate.state_vector_call(initial_rho) # pylint: disable=no-member
+    qibo.set_backend(original_backend)
+
+
+@pytest.mark.parametrize("t1,t2,time,excpop",
+                         [(1.0, 0.5, 1.5, 1.5), (1.0, 0.5, -0.5, 0.5),
+                          (1.0, -0.5, 1.5, 0.5), (-1.0, 0.5, 1.5, 0.5),
+                          (1.0, 3.0, 1.5, 0.5)])
+def test_thermal_relaxation_channel_errors(backend, t1, t2, time, excpop):
+    original_backend = qibo.get_backend()
+    qibo.set_backend(backend)
+    with pytest.raises(ValueError):
+        gate = gates.ThermalRelaxationChannel(
+            0, t1, t2, time, excited_population=excpop)
+    qibo.set_backend(original_backend)
